perf: cache admin key report notifications
